Split PolymorphicListAsMapSerializer into a helper class CustomMapSerializer
